+++ src/app/features/events/events-search-dialog/events-search-dialog.component.ts
@@ -40,7 +40,7 @@ export class EventsSearchDialogComponent implements OnInit, OnDestroy {
   ngOnInit(): void {
 
     const subs1$ =
-      this.eventSrv.getAllEventsWithAppointments(false)
+      this.eventSrv.getAllEventsWithAppointments(true, false)
         .pipe(
           map(events => events.map(event => {
             const reducer = (acc, value) => `${acc} ${value.substr(0, value.indexOf(' '))}`;